PDDA-385 Removing the unneeded abstract layer to store blob and moving it to a transient field
