--- src/main/java/uk/gov/hmcts/pdda/business/entities/xhbcourtellist/AbstractCourtelDao.java
@@ -1,19 +1,0 @@
-package uk.gov.hmcts.pdda.business.entities.xhbcourtellist;
-
-import uk.gov.hmcts.pdda.business.entities.AbstractVersionedDao;
-import uk.gov.hmcts.pdda.business.entities.xhbblob.XhbBlobDao;
-
-public abstract class AbstractCourtelDao extends AbstractVersionedDao {
-
-    private static final long serialVersionUID = 1L;
-    
-    private XhbBlobDao blob;
-
-    public XhbBlobDao getBlob() {
-        return blob;
-    }
-
-    public void setBlob(XhbBlobDao blob) {
-        this.blob = blob;
-    }
-}
